Hexmapper

+++ src/perlin.cpp
@@ -48,7 +48,7 @@ DRAW_IMAGE(draw_image)
                     //n = random_bilateral(&randomize);
                 }
                 
-                u32 grayScale = round(255.0f * (0.5f * n + 0.5f)) & 0xFF;
+                u32 grayScale = (u32)round(255.0f * (0.5f * n + 0.5f)) & 0xFF;
                 
             image->pixels[y * image->width + x] = 
                     (0xFF << 24) | (grayScale << 16) | (grayScale << 8) | grayScale;
